Add Tailwind and other CSS to reader package
I installed Tailwind, and imported Google Fonts for package reader.

The package was not self-contained in terms of CSS styles, which
made the styles the project that imported reader package trickle down
into the Reader UI, which made its style unpredictable.

To make the reader package self-contained, I installed Tailwindcss, and
the Google Fonts and symbols, which the reader UI requires.