Turn tooltip into modal for navigating btw statements
Also add "View code" button and render code in code mirror component

Overall before it wasn't quite sure if this is code or something else.
Also wasn't possible to navigate the codebase, since I disabled links.

Now it is a lot more clear that this project is code. And now it's possible
to navigate like one. Overall, better UX.